combine test data files
